[REFACTOR] : DTO이름 변경
