+++ be/sidedish-be/src/main/java/com/codesquad/sidedish/web/sidedish/service/SidedishCategoryService.java
@@ -1,7 +1,7 @@
 package com.codesquad.sidedish.web.sidedish.service;
 
-import com.codesquad.sidedish.web.sidedish.DTO.ItemDTO;
 import com.codesquad.sidedish.web.sidedish.DTO.SidedishDTO;
+import com.codesquad.sidedish.web.sidedish.DTO.BestSidedishDTO;
 import com.codesquad.sidedish.web.sidedish.domain.SidedishCategory;
 import com.codesquad.sidedish.web.sidedish.repository.SidedishCategoryRepository;
 import org.springframework.stereotype.Service;
@@ -19,15 +19,15 @@ public class SidedishCategoryService {
         this.imageService = imageService;
     }
 
-    public List<SidedishDTO> readBestCategories() {
+    public List<BestSidedishDTO> readBestCategories() {
         return sidedishCategoryRepository.findByIsBestTrue().stream()
-                .map(sidedishCategory -> SidedishDTO.of(sidedishCategory, imageService.readImages()))
+                .map(sidedishCategory -> BestSidedishDTO.of(sidedishCategory, imageService.readImages()))
                 .collect(Collectors.toList());
     }
 
-    public List<ItemDTO> readSideSidedishesBy(String categoryName) {
+    public List<SidedishDTO> readSideSidedishesBy(String categoryName) {
         return readBy(categoryName).sidedishStream()
-                .map(sidedish -> ItemDTO.of(sidedish, imageService.readImages()))
+                .map(sidedish -> SidedishDTO.of(sidedish, imageService.readImages()))
                 .collect(Collectors.toList());
     }
